refactor: extract DaemonState, config, plugins, and init helpers from daemon/mod.rs
Move DaemonConfig, PrCoworkerCache, and DaemonState to src/daemon/state.rs.
Move plugin management (ensure_plugins_installed, get_installed_plugins,
install_plugin) to src/daemon/plugins.rs. Move init helpers
(load_additional_worktree_managers, acquire_pid_lock) to src/daemon/setup.rs.

Rename misleading constants:
- IDLE_CHECK_INTERVAL → SESSION_MONITOR_INTERVAL
- ORPHAN_CHECK_INTERVAL_SECS → TASK_DISPATCH_INTERVAL_SECS
- idle_check_interval → session_monitor_interval (variable)
- orphan_check_interval → task_dispatch_interval (variable)

mod.rs now contains only: module declarations, re-exports, run() event loop,
and tests. Reduces mod.rs from ~2300 lines to ~1675 lines.
